Refactor: mueve archivos de navegación a ui.navigation
