+++ app/src/main/java/ar/edu/unlam/scaffoldingandroid3/ui/navigation/NavigationGraph.kt
@@ -1,4 +1,4 @@
-package ar.edu.unlam.scaffoldingandroid3.navigation
+package ar.edu.unlam.scaffoldingandroid3.ui.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
