+++ app/src/main/java/ar/edu/unlam/scaffoldingandroid3/ui/navigation/Screen.kt
@@ -1,4 +1,4 @@
-package ar.edu.unlam.scaffoldingandroid3.navigation
+package ar.edu.unlam.scaffoldingandroid3.ui.navigation
 
 /**
  * TODO: Sealed class - Destinos de navegación
